Crate asset_hub_polkadot_runtime

Asset Hub Polkadot Runtime

Asset Hub Polkadot is a parachain that provides an interface to create, manage, and use assets. Assets may be fungible or non-fungible.

Renaming

This chain was originally known as “Statemint”. You may see references to Statemint, Statemine, and Westmint throughout the codebase. These are synonymous with “Asset Hub Polkadot, Kusama, and Westend”, respectively.

Assets

Fungibles: Configuration of pallet-assets.

Non-Fungibles (NFTs): Configuration of pallet-uniques.

Other Functionality

Native Balances

Asset Hub Polkadot uses its parent DOT token as its native asset.

Governance

As a system parachain, Asset Hub defers its governance (namely, its Root origin), to its Relay Chain parent, Polkadot.

Collator Selection

Asset Hub uses pallet-collator-selection, a simple first-come-first-served registration system where collators can reserve a small bond to join the block producer set. There is no slashing.

XCM

Because Asset Hub is fully under the control of the Relay Chain, it is meant to be a TrustedTeleporter. It can also serve as a reserve location to other parachains for DOT as well as other local assets.

Re-exports

pub use parachains_common as common;

Modules

api

constants

xcm_config

Structs

AnnouncementDepositBase

AnnouncementDepositFactor

ApprovalDeposit

AssetAccountDeposit

AssetDeposit

AssetsStringLimit

DepositBase

DepositFactor

ExistentialDeposit

FellowsBodyId

ForeignAssetsApprovalDeposit

ForeignAssetsAssetAccountDeposit

ForeignAssetsAssetDeposit

ForeignAssetsAssetsStringLimit

ForeignAssetsMetadataDepositBase

ForeignAssetsMetadataDepositPerByte

MaxPending

MaxProxies

MaxSignatories

MetadataDepositBase

Key = 32 bytes, Value = 36 bytes (32+1+1+1+1)

MetadataDepositPerByte

NftsAttributeDepositBase

NftsCollectionDeposit

NftsDepositPerByte

NftsItemDeposit

NftsMaxDeadlineDuration

NftsMetadataDepositBase

NftsPalletFeatures

Offset

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Period

PotId

ProxyDepositBase

ProxyDepositFactor

ReservedDmpWeight

ReservedXcmpWeight

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SessionKeys

SessionLength

StakingAdminBodyId

TransactionByteFee

Relay Chain TransactionByteFee / 10

UniquesAttributeDepositBase

UniquesCollectionDeposit

UniquesDepositPerByte

UniquesItemDeposit

UniquesMetadataDepositBase

Version

Enums

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

Constants

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

AssetTxPayment

Assets

AssetsForceOrigin

We allow root to execute privileged asset operations.

Aura

AuraConfig

AuraExt

AuraExtConfig

Authorship

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

CollatorSelection

CollatorSelectionConfig

CollatorSelectionUpdateOrigin

We allow root and the StakingAdmin to execute privileged collator selection operations.

CumulusXcm

DmpQueue

Executive

Executive: handles dispatch to the various modules.

ForeignAssets

ForeignAssetsInstance

Assets managed by some foreign location. Note: we do not declare a ForeignAssetsCall type, as this type is used in proxy definitions. We assume that a foreign location would not want to set an individual, local account as a proxy for the issuance of their assets. This issuance should be managed by the foreign location’s governance.

GenesisConfigDeprecated

Migrations

Migrations to apply on runtime upgrade.

Multisig

Nfts

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

Proxy

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

System

SystemConfig

Timestamp

TransactionPayment

TrustBackedAssetsInstance

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Uniques

Utility

XcmpQueue