Module asset_hub_polkadot_runtime::xcm_config

Structs

CheckingAccount

DotLocation

FellowsPlurality

FellowshipLocation

FellowshipSalaryPallet

GovernanceLocation

MaxAssetsIntoHolding

MaxInstructions

ParentOrParentsPlurality

ParentOrSiblings

RelayChainOrigin

RelayNetwork

SafeCallFilter

A call filter for the XCM Transact instruction. This is a temporary measure until we properly account for proof size weights.

TrustBackedAssetsPalletLocation

UniversalLocation

UniversalLocationNetworkId

XcmAssetFeesReceiver

XcmBenchmarkHelper

Simple conversion of u32 into an AssetId for use in benchmarking.

XcmConfig

Type Aliases

AssetFeeAsExistentialDepositMultiplierFeeCharger

AssetTransactors

Means for transacting assets on this chain.

Barrier

CurrencyTransactor

Means for transacting the native currency on this chain.

ForeignAssetsConvertedConcreteId

AssetId/Balance converter for TrustBackedAssets

ForeignCreatorsSovereignAccountOf

ForeignFungiblesTransactor

Means for transacting foreign assets from different global consensus.

FungiblesTransactor

Means for transacting assets besides the native currency on this chain.

LocalOriginToLocation

Converts a local signed origin into an XCM multilocation. Forms the basis for local origins sending/executing XCMs.

LocationToAccountId

Type for specifying how a MultiLocation can be converted into an AccountId. This is used when determining ownership of accounts for asset transacting and when attempting to use XCM Transact in order to determine the dispatch Origin.

TrustBackedAssetsConvertedConcreteId

AssetId/Balance converter for TrustBackedAssets.

XcmOriginToTransactDispatchOrigin

This is the type we use to convert an (incoming) XCM origin into a local Origin instance, ready for dispatching a transaction with Xcm’s Transact. There is an OriginKind which can biases the kind of local Origin it will become.

XcmRouter

The means for routing XCM messages which are not for local execution into the right message queues.